requirements.txt
pandas
faker
boto3
python-dotenv
prefect